+++ packages/common/src/ScrollResponder.ts
@@ -17,7 +17,7 @@ export class ScrollResponder {
     private execFunc: ScrollRequestHandler,
     private emitter: Emitter<CalendarListeners>,
     private scrollTime: Duration,
-    private preserveScroll: boolean,
+    private scrollTimeReset: boolean,
   ) {
     emitter.on('_scrollRequest', this.handleScrollRequest)
     this.fireInitialScroll()
@@ -28,7 +28,7 @@ export class ScrollResponder {
   }
 
   update(isDatesNew: boolean) {
-    if (isDatesNew && !this.preserveScroll) {
+    if (isDatesNew && this.scrollTimeReset) {
       this.fireInitialScroll() // will drain
     } else {
       this.drain()